Remove partial types, pass tests
