+++ weld/ast.rs
@@ -1,11 +1,290 @@
-//! Abstract syntax tree for Weld.
+//! Defines types in the Weld IR.
 
-use std::vec;
-use std::fmt;
-use std::hash::Hash;
-
+use super::annotations::Annotations;
 use super::error::*;
-use super::annotations::*;
+
+use self::Type::*;
+use self::ExprKind::*;
+use self::ScalarKind::*;
+use self::BuilderKind::*;
+use self::BinOpKind::*;
+
+use std::fmt;
+use std::vec;
+
+#[derive(Clone, Debug, PartialEq, Eq, Hash)]
+pub enum Type {
+    /// A scalar.
+    Scalar(ScalarKind),
+    /// A SIMD vector.
+    Simd(ScalarKind),
+    /// A variable-length vector.
+    Vector(Box<Type>),
+    /// A dictionary mapping keys to values.
+    Dict(Box<Type>, Box<Type>),
+    /// A mutable builder to construct results.
+    Builder(BuilderKind, Annotations),
+    /// An ordered struct or tuple.
+    Struct(Vec<Type>),
+    /// A function with a list of arguments and return type.
+    Function(Vec<Type>, Box<Type>),
+    /// An unknown type, used only for type inference.
+    Unknown,
+}
+
+impl Type {
+    /// Returns the child types of this `Type`.
+    pub fn children(&self) -> vec::IntoIter<&Type> {
+        match *self {
+            Unknown | Scalar(_) | Simd(_) => vec![],
+            Vector(ref elem) => {
+                vec![elem.as_ref()]
+            }
+            Dict(ref key, ref value) => {
+                vec![key.as_ref(), value.as_ref()]
+            }
+            Builder(ref kind, _) => match kind {
+                Appender(ref elem) => {
+                    vec![elem.as_ref()]
+                }
+                Merger(ref elem, _) => {
+                    vec![elem.as_ref()]
+                }
+                DictMerger(ref key, ref value, _) => {
+                    vec![key.as_ref(), value.as_ref()]
+                }
+                GroupMerger(ref key, ref value) =>  {
+                    vec![key.as_ref(), value.as_ref()]
+                }
+                VecMerger(ref elem, _) => {
+                    vec![elem.as_ref()]
+                }
+            },
+            Struct(ref elems) => elems.iter().collect(),
+            Function(ref params, ref res) => {
+                let mut children = vec![];
+                for param in params.iter() {
+                    children.push(param);
+                }
+                children.push(res.as_ref());
+                children
+            }
+        }.into_iter()
+    }
+
+    /// Returns whether this `Type` is a SIMD value.
+    ///
+    /// A value is a SIMD value if its a `Simd` type or it is a `Struct` where each member is a
+    /// `Simd` type. We additionally consider each of the builders to be SIMD values, since they
+    /// can operate over SIMD values as inputs.
+    pub fn is_simd(&self) -> bool {
+        match *self {
+            Simd(_) | Builder(_, _) => true,
+            Struct(ref fields) => fields.iter().all(|f| f.is_simd()),
+            _ => false
+        }
+    }
+
+    pub fn is_scalar(&self) -> bool {
+        match *self {
+            Scalar(_) => true,
+            _ => false
+        }
+    }
+
+    /// Return the vectorized version of a type.
+    ///
+    /// This method returns an error if this `Type` is not vectorizable.
+    pub fn simd_type(&self) -> WeldResult<Type> {
+        match *self {
+            Scalar(kind) => Ok(Simd(kind)),
+            Builder(_, _) => Ok(self.clone()),
+            Struct(ref fields) => {
+                let result: WeldResult<_> = fields.iter().map(|f| f.simd_type()).collect();
+                Ok(Struct(result?))
+            }
+            _ => compile_err!("simd_type called on non-SIMD {:?}", self)
+        }
+    }
+
+    /// Return the scalar version of a type.
+    ///
+    /// This method returns an error if this `Type` is not scalarizable.
+    pub fn scalar_type(&self) -> WeldResult<Type> {
+        match *self {
+            Simd(kind) => Ok(Scalar(kind)),
+            Builder(_, _) => Ok(self.clone()),
+            Struct(ref fields) => {
+                let result: WeldResult<_> = fields.iter().map(|f| f.scalar_type()).collect();
+                Ok(Struct(result?))
+            }
+            _ => compile_err!("scalar_type called on non-SIMD {:?}", self)
+        }
+    }
+
+    /// Returns the type merged into a builder.
+    ///
+    /// Returns an error if this `Type` is not a builder type.
+    pub fn merge_type(&self) -> WeldResult<Type> {
+        if let Builder(ref kind, _) = *self {
+            Ok(kind.merge_type())
+
+        } else {
+            compile_err!("merge_type called on non-builder type {:?}", self)
+        }
+    }
+
+    /// Returns whether this `Type` is partial.
+    ///
+    /// A type is partial if it or any of its subtypes is `Unknown`.
+    pub fn partial_type(&self) -> bool {
+        match *self {
+            Unknown => true,
+            _ => self.children().any(|t| t.partial_type()),
+        }
+    }
+}
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
+/// Scalar types in the Weld IR.
+pub enum ScalarKind {
+    Bool,
+    I8,
+    I16,
+    I32,
+    I64,
+    U8,
+    U16,
+    U32,
+    U64,
+    F32,
+    F64,
+}
+
+impl ScalarKind {
+    /// Returns whether this scalar is a floating-point type.
+    ///
+    /// The current floating point kinds are `F32` and `F64`.
+    pub fn is_float(&self) -> bool {
+        match *self {
+            F32 | F64 => true,
+            _ => false
+        }
+    }
+
+    /// Returns whether this scalar is a boolean.
+    pub fn is_bool(&self) -> bool {
+        match *self {
+            Bool => true,
+            _ => false
+        }
+    }
+
+    /// Returns whether this scalar is a signed integer.
+    ///
+    /// Booleans are not considered to be signed integers.
+    pub fn is_signed_integer(&self) -> bool {
+        match *self {
+            I8 | I16 | I32 | I64 => true,
+            _ => false
+        }
+    }
+
+    /// Returns whether this scalar is an unsigned integer.
+    ///
+    /// Booleans are not considered to be unsigned integers.
+    pub fn is_unsigned_integer(&self) -> bool {
+        match *self {
+            U8 | U16 | U32 | U64 => true,
+            _ => false
+        }
+    }
+
+    /// Returns whether this scalar is an integer.
+    ///
+    /// Booleans are not considered to be integers.
+    pub fn is_integer(&self) -> bool {
+        return self.is_signed_integer() || self.is_unsigned_integer();
+    }
+
+    /// Return the length of this scalar type in bits.
+    pub fn bits(&self) -> u32 {
+        match *self {
+            Bool => 1,
+            I8 | U8 => 8,
+            I16 | U16 => 16,
+            I32 | U32 | F32 => 32,
+            I64 | U64 | F64 => 64
+        }
+    }
+}
+
+impl fmt::Display for ScalarKind {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        let text = match *self {
+            Bool => "bool",
+            I8 => "i8",
+            I16 => "i16",
+            I32 => "i32",
+            I64 => "i64",
+            U8 => "u8",
+            U16 => "u16",
+            U32 => "u32",
+            U64 => "u64",
+            F32 => "f32",
+            F64 => "f64",
+        };
+        f.write_str(text)
+    }
+}
+
+#[derive(Clone, Debug, PartialEq, Eq, Hash)]
+pub enum BuilderKind {
+    /// A builder that appends items to a list.
+    Appender(Box<Type>),
+    /// A builder that constructs a scalar or struct.
+    ///
+    /// Results are merged using an associative binary operator.
+    Merger(Box<Type>, BinOpKind),
+    /// A builder that creates a dictionary.
+    ///
+    /// Results are grouped by key, and values are combined using an associative binary operator.
+    DictMerger(Box<Type>, Box<Type>, BinOpKind),
+    /// A builder that creates groups.
+    ///
+    /// Results are grouped by key, where values are appended to a list.
+    GroupMerger(Box<Type>, Box<Type>),
+    /// A builder that constructs a vector by updating elements.
+    ///
+    /// Elements are updated by index using an associative binary operator.
+    VecMerger(Box<Type>, BinOpKind),
+}
+
+impl BuilderKind {
+    /// Returns the type merged into this `BuilderKind`.
+    pub fn merge_type(&self) -> Type {
+        match *self {
+            Appender(ref elem) => *elem.clone(),
+            Merger(ref elem, _) => *elem.clone(),
+            DictMerger(ref key, ref value, _) => Struct(vec![*key.clone(), *value.clone()]),
+            GroupMerger(ref key, ref value) => Struct(vec![*key.clone(), *value.clone()]),
+            VecMerger(ref elem, _) => Struct(vec![Scalar(I64), *elem.clone()]),
+        }
+    }
+
+    /// Returns the type produced by this `BuilderKind`.
+    pub fn result_type(&self) -> Type {
+        match *self {
+            Appender(ref elem) => Vector(elem.clone()),
+            Merger(ref elem, _) => *elem.clone(),
+            DictMerger(ref key, ref value, _) => Dict(key.clone(), value.clone()),
+            GroupMerger(ref key, ref value) => Dict(key.clone(), Box::new(Vector(value.clone()))),
+            VecMerger(ref elem, _) => Vector(elem.clone()),
+        }
+    }
+}
+
+// -------------------------------
 
 /// A symbol (identifier name); for now these are strings, but we may add some kind of scope ID.
 #[derive(Clone, Debug, PartialEq, Eq, Hash, PartialOrd, Ord)]
@@ -40,195 +319,13 @@ impl fmt::Display for Symbol {
     }
 }
 
-/// A data type.
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub enum Type {
-    Scalar(ScalarKind),
-    Simd(ScalarKind),
-    Vector(Box<Type>),
-    Dict(Box<Type>, Box<Type>),
-    Builder(BuilderKind, Annotations),
-    Struct(Vec<Type>),
-    Function(Vec<Type>, Box<Type>),
-}
-
-impl Type {
-    /// Return true if a given type is SIMD vectorized, which concretely means that it is
-    /// either a simd[T], a builder, or a struct of SIMD types.
-    ///
-    /// TODO: we should investigate just using simd[struct[T]] for structs and turning it
-    /// to the nested form at runtime.
-    pub fn is_simd(&self) -> bool {
-        use self::Type::*;
-        match *self {
-            Simd(_) => true,
-            Builder(_, _) => true,
-            Struct(ref fields) => fields.iter().all(|f| f.is_simd()),
-            _ => false
-        }
-    }
-
-    pub fn is_scalar(&self) -> bool {
-        use self::Type::*;
-        match *self {
-            Scalar(_) => true,
-            _ => false
-        }
-    }
-
-    /// Get the vectorized version of a type, if it is vectorizable.
-    pub fn simd_type(&self) -> WeldResult<Type> {
-        use self::Type::*;
-        match *self {
-            Scalar(kind) => Ok(Simd(kind)),
-
-            Builder(_, _) => Ok(self.clone()),
-
-            Struct(ref fields) => {
-                let mut new_fields = vec![];
-                for f in fields {
-                    new_fields.push(f.simd_type()?)
-                }
-                Ok(Struct(new_fields))
-            }
-
-            _ => compile_err!("simd_type called on non-SIMD {:?}", self)
-        }
-    }
-
-    /// Get the scalar version of a type if it `is_simd`.
-    pub fn scalar_type(&self) -> WeldResult<Type> {
-        use self::Type::*;
-        match *self {
-            Simd(kind) => Ok(Scalar(kind)),
-
-            Builder(_, _) => Ok(self.clone()),
-
-            Struct(ref fields) => {
-                let mut new_fields = vec![];
-                for f in fields {
-                    new_fields.push(f.scalar_type()?)
-                }
-                Ok(Struct(new_fields))
-            }
-
-            _ => compile_err!("scalar_type called on non-SIMD {:?}", self)
-        }
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
-pub enum ScalarKind {
-    Bool,
-    I8,
-    I16,
-    I32,
-    I64,
-    U8,
-    U16,
-    U32,
-    U64,
-    F32,
-    F64,
-}
-
-impl ScalarKind {
-    /// Is this scalar a floating point type (F32 or F64)?
-    pub fn is_float(&self) -> bool {
-        use ast::ScalarKind::*;
-        match *self {
-            F32 | F64 => true,
-            _ => false
-        }
-    }
-
-    /// Is this scalar a Bool?
-    pub fn is_bool(&self) -> bool {
-        use ast::ScalarKind::*;
-        match *self {
-            Bool => true,
-            _ => false
-        }
-    }
-
-    /// Is this type a signed integer of any type (excludes Bool)?
-    pub fn is_signed_integer(&self) -> bool {
-        use ast::ScalarKind::*;
-        match *self {
-            I8 | I16 | I32 | I64 => true,
-            _ => false
-        }
-    }
-
-    /// Is this type a signed integer of any type (excludes Bool)?
-    pub fn is_unsigned_integer(&self) -> bool {
-        use ast::ScalarKind::*;
-        match *self {
-            U8 | U16 | U32 | U64 => true,
-            _ => false
-        }
-    }
-
-    /// Is this type a signed or unsigned integer of any type (excludes Bool)?
-    pub fn is_integer(&self) -> bool {
-        return self.is_signed_integer() || self.is_unsigned_integer();
-    }
-
-    /// Return the length of this scalar type in bits
-    pub fn bits(&self) -> u32 {
-        use ast::ScalarKind::*;
-        match *self {
-            Bool => 1,
-            I8 | U8 => 8,
-            I16 | U16 => 16,
-            I32 | U32 | F32 => 32,
-            I64 | U64 | F64 => 64
-        }
-    }
-}
-
-impl fmt::Display for ScalarKind {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        use ast::ScalarKind::*;
-        let text = match *self {
-            Bool => "bool",
-            I8 => "i8",
-            I16 => "i16",
-            I32 => "i32",
-            I64 => "i64",
-            U8 => "u8",
-            U16 => "u16",
-            U32 => "u32",
-            U64 => "u64",
-            F32 => "f32",
-            F64 => "f64",
-        };
-        f.write_str(text)
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub enum BuilderKind {
-    Appender(Box<Type>),
-    Merger(Box<Type>, BinOpKind),
-    /// key_type, value_type, binop
-    DictMerger(Box<Type>, Box<Type>, BinOpKind),
-    GroupMerger(Box<Type>, Box<Type>),
-    /// elem_type, binop
-    VecMerger(Box<Type>, BinOpKind),
-}
-
-pub trait TypeBounds: Clone + PartialEq + Hash {}
-
-impl TypeBounds for Type {}
-
 /// An expression tree, having type annotations of type T. We make this parametrized because
 /// expressions have different "kinds" of types attached to them at different points in the
 /// compilation process -- namely PartialType when parsed and then Type after type inference.
 #[derive(Clone, Debug, PartialEq)]
-pub struct Expr<T: TypeBounds> {
-    pub ty: T,
-    pub kind: ExprKind<T>,
+pub struct Expr {
+    pub ty: Type,
+    pub kind: ExprKind,
     pub annotations: Annotations,
 }
 
@@ -244,7 +341,7 @@ pub enum IterKind {
 
 impl fmt::Display for IterKind {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        use ast::IterKind::*;
+        use self::IterKind::*;
         let text = match *self {
             ScalarIter => "scalar",
             SimdIter => "vectorized",
@@ -259,18 +356,17 @@ impl fmt::Display for IterKind {
 /// An iterator, which specifies a vector to iterate over and optionally a start index,
 /// end index, and stride.
 #[derive(Clone, Debug, PartialEq)]
-pub struct Iter<T: TypeBounds> {
-    pub data: Box<Expr<T>>,
-    pub start: Option<Box<Expr<T>>>,
-    pub end: Option<Box<Expr<T>>>,
-    pub stride: Option<Box<Expr<T>>>,
+pub struct Iter {
+    pub data: Box<Expr>,
+    pub start: Option<Box<Expr>>,
+    pub end: Option<Box<Expr>>,
+    pub stride: Option<Box<Expr>>,
     pub kind: IterKind,
-    // NdIter specific fields
-    pub strides: Option<Box<Expr<T>>>,
-    pub shape: Option<Box<Expr<T>>>,
+    pub strides: Option<Box<Expr>>,
+    pub shape: Option<Box<Expr>>,
 }
 
-impl<T: TypeBounds> Iter<T> {
+impl Iter {
     /// Returns true if this is a simple iterator with no start/stride/end specified
     /// (i.e., it iterates over all the input data) and kind `ScalarIter`.
     pub fn is_simple(&self) -> bool {
@@ -280,104 +376,101 @@ impl<T: TypeBounds> Iter<T> {
 }
 
 #[derive(Clone, Debug, PartialEq)]
-pub enum ExprKind<T: TypeBounds> {
-    // TODO: maybe all of these should take named parameters
+pub enum ExprKind {
     Literal(LiteralKind),
     Ident(Symbol),
-    Negate(Box<Expr<T>>),
+    Negate(Box<Expr>),
     // Broadcasts a scalar into a vector, e.g., 1 -> <1, 1, 1, 1>
-    Broadcast(Box<Expr<T>>),
+    Broadcast(Box<Expr>),
     BinOp {
         kind: BinOpKind,
-        left: Box<Expr<T>>,
-        right: Box<Expr<T>>,
+        left: Box<Expr>,
+        right: Box<Expr>,
     },
     UnaryOp {
         kind: UnaryOpKind,
-        value: Box<Expr<T>>,
+        value: Box<Expr>,
     },
     Cast {
         kind: ScalarKind,
-        child_expr: Box<Expr<T>>,
+        child_expr: Box<Expr>,
     },
-    ToVec { child_expr: Box<Expr<T>> },
-    MakeStruct { elems: Vec<Expr<T>> },
-    MakeVector { elems: Vec<Expr<T>> },
-    Zip { vectors: Vec<Expr<T>> },
-    GetField { expr: Box<Expr<T>>, index: u32 },
-    Length { data: Box<Expr<T>> },
+    ToVec { child_expr: Box<Expr> },
+    MakeStruct { elems: Vec<Expr> },
+    MakeVector { elems: Vec<Expr> },
+    Zip { vectors: Vec<Expr> },
+    GetField { expr: Box<Expr>, index: u32 },
+    Length { data: Box<Expr> },
     Lookup {
-        data: Box<Expr<T>>,
-        index: Box<Expr<T>>,
+        data: Box<Expr>,
+        index: Box<Expr>,
     },
     KeyExists {
-        data: Box<Expr<T>>,
-        key: Box<Expr<T>>,
+        data: Box<Expr>,
+        key: Box<Expr>,
     },
     Slice {
-        data: Box<Expr<T>>,
-        index: Box<Expr<T>>,
-        size: Box<Expr<T>>,
+        data: Box<Expr>,
+        index: Box<Expr>,
+        size: Box<Expr>,
     },
     Sort {
-        data: Box<Expr<T>>,
-        keyfunc: Box<Expr<T>>,
+        data: Box<Expr>,
+        keyfunc: Box<Expr>,
     },
     Let {
         name: Symbol,
-        value: Box<Expr<T>>,
-        body: Box<Expr<T>>,
+        value: Box<Expr>,
+        body: Box<Expr>,
     },
     If {
-        cond: Box<Expr<T>>,
-        on_true: Box<Expr<T>>,
-        on_false: Box<Expr<T>>,
+        cond: Box<Expr>,
+        on_true: Box<Expr>,
+        on_false: Box<Expr>,
     },
-    /// Sequential loop
     Iterate {
-        initial: Box<Expr<T>>,
-        update_func: Box<Expr<T>>,
+        initial: Box<Expr>,
+        update_func: Box<Expr>,
     },
     Select {
-        cond: Box<Expr<T>>,
-        on_true: Box<Expr<T>>,
-        on_false: Box<Expr<T>>,
+        cond: Box<Expr>,
+        on_true: Box<Expr>,
+        on_false: Box<Expr>,
     },
     Lambda {
-        params: Vec<Parameter<T>>,
-        body: Box<Expr<T>>,
+        params: Vec<Parameter>,
+        body: Box<Expr>,
     },
     Apply {
-        func: Box<Expr<T>>,
-        params: Vec<Expr<T>>,
+        func: Box<Expr>,
+        params: Vec<Expr>,
     },
     CUDF {
         sym_name: String,
-        args: Vec<Expr<T>>,
-        return_ty: Box<T>,
+        args: Vec<Expr>,
+        return_ty: Box<Type>,
     },
-    Serialize(Box<Expr<T>>),
+    Serialize(Box<Expr>),
     Deserialize {
-        value: Box<Expr<T>>,
-        value_ty: Box<T>,
+        value: Box<Expr>,
+        value_ty: Box<Type>,
     },
-    NewBuilder(Option<Box<Expr<T>>>),
+    NewBuilder(Option<Box<Expr>>),
     For {
-        iters: Vec<Iter<T>>,
-        builder: Box<Expr<T>>,
-        func: Box<Expr<T>>,
+        iters: Vec<Iter>,
+        builder: Box<Expr>,
+        func: Box<Expr>,
     },
     Merge {
-        builder: Box<Expr<T>>,
-        value: Box<Expr<T>>,
+        builder: Box<Expr>,
+        value: Box<Expr>,
     },
-    Res { builder: Box<Expr<T>> },
+    Res { builder: Box<Expr> },
 }
 
-impl<T: TypeBounds> ExprKind<T> {
+impl ExprKind {
     /// Return a readable name for the kind which is independent of any subexpressions.
     pub fn name(&self) -> &str {
-        use ast::ExprKind::*;
         match *self {
             Literal(_) => "Literal",
             Ident(_) => "Ident",
@@ -413,7 +506,6 @@ impl<T: TypeBounds> ExprKind<T> {
     }
 
     pub fn is_builder_expr(&self) -> bool {
-        use ast::ExprKind::*;
         match *self {
             Merge { .. } | Res { .. } | For { .. } | NewBuilder(_) => true,
             _ => false,
@@ -432,9 +524,7 @@ pub enum LiteralKind {
     U16Literal(u16),
     U32Literal(u32),
     U64Literal(u64),
-    // stored as raw bits.
     F32Literal(u32),
-    // stored as raw bits.
     F64Literal(u64),
     StringLiteral(String),
 }
@@ -481,7 +571,6 @@ pub enum UnaryOpKind {
 
 impl BinOpKind {
     pub fn is_comparison(&self) -> bool {
-        use ast::BinOpKind::*;
         match *self {
             Equal | NotEqual | LessThan | GreaterThan | LessThanOrEqual | GreaterThanOrEqual => {
                 true
@@ -493,7 +582,6 @@ impl BinOpKind {
 
 impl fmt::Display for BinOpKind {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        use ast::BinOpKind::*;
         let text = match *self {
             Add => "+",
             Subtract => "-",
@@ -527,23 +615,14 @@ impl fmt::Display for UnaryOpKind {
 }
 
 #[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub struct Parameter<T: TypeBounds> {
+pub struct Parameter {
     pub name: Symbol,
-    pub ty: T,
+    pub ty: Type,
 }
 
-/// A typed expression struct.
-pub type TypedExpr = Expr<Type>;
-
-/// A typed iterator.
-pub type TypedIter = Iter<Type>;
- 
-/// A typed parameter.
-pub type TypedParameter = Parameter<Type>;
-
-impl<T: TypeBounds> Expr<T> {
+impl Expr {
     /// Get an iterator for the children of this expression.
-    pub fn children(&self) -> vec::IntoIter<&Expr<T>> {
+    pub fn children(&self) -> vec::IntoIter<&Expr> {
         use self::ExprKind::*;
         match self.kind {
             BinOp {
@@ -589,7 +668,7 @@ impl<T: TypeBounds> Expr<T> {
                 ref builder,
                 ref func,
             } => {
-                let mut res: Vec<&Expr<T>> = vec![];
+                let mut res: Vec<&Expr> = vec![];
                 for iter in iters {
                     res.push(iter.data.as_ref());
                     if let Some(ref s) = iter.start {
@@ -646,7 +725,7 @@ impl<T: TypeBounds> Expr<T> {
     }
 
     /// Get an iterator of mutable references to the children of this expression.
-    pub fn children_mut(&mut self) -> vec::IntoIter<&mut Expr<T>> {
+    pub fn children_mut(&mut self) -> vec::IntoIter<&mut Expr> {
         use self::ExprKind::*;
         match self.kind {
             BinOp {
@@ -695,7 +774,7 @@ impl<T: TypeBounds> Expr<T> {
                 ref mut builder,
                 ref mut func,
             } => {
-                let mut res: Vec<&mut Expr<T>> = vec![];
+                let mut res: Vec<&mut Expr> = vec![];
                 for iter in iters {
                     res.push(iter.data.as_mut());
                     if let Some(ref mut s) = iter.start {
@@ -756,17 +835,17 @@ impl<T: TypeBounds> Expr<T> {
     /// Symbols in the two expressions must have a one to one correspondance for the trees to be
     /// considered equal. If an undefined symbol is encountered during the comparison, it must
     /// be the same in both expressions (e.g. for a symbol captured from an outer scope).
-    pub fn compare_ignoring_symbols(&self, other: &Expr<T>) -> WeldResult<bool> {
+    pub fn compare_ignoring_symbols(&self, other: &Expr) -> WeldResult<bool> {
         use self::ExprKind::*;
         use std::collections::HashMap;
         let mut sym_map: HashMap<&Symbol, &Symbol> = HashMap::new();
         let mut reverse_sym_map: HashMap<&Symbol, &Symbol> = HashMap::new();
 
-        fn _compare_ignoring_symbols<'b, 'a, U: TypeBounds>(e1: &'a Expr<U>,
-                                                            e2: &'b Expr<U>,
-                                                            sym_map: &mut HashMap<&'a Symbol, &'b Symbol>,
-                                                            reverse_sym_map: &mut HashMap<&'b Symbol, &'a Symbol>)
-                                                            -> WeldResult<bool> {
+        fn _compare_ignoring_symbols<'b, 'a>(e1: &'a Expr,
+                                             e2: &'b Expr,
+                                             sym_map: &mut HashMap<&'a Symbol, &'b Symbol>,
+                                             reverse_sym_map: &mut HashMap<&'b Symbol, &'a Symbol>)
+            -> WeldResult<bool> {
             // First, check the type.
             if e1.ty != e2.ty {
                 return Ok(false);
@@ -877,7 +956,7 @@ impl<T: TypeBounds> Expr<T> {
 
     /// Substitute Ident nodes with the given symbol for another expression, stopping when an
     /// expression in the tree redefines the symbol (e.g. Let or Lambda parameters).
-    pub fn substitute(&mut self, symbol: &Symbol, replacement: &Expr<T>) {
+    pub fn substitute(&mut self, symbol: &Symbol, replacement: &Expr) {
         // Replace ourselves if we are exactly the symbol.
         use self::ExprKind::*;
         let mut self_matches = false;
@@ -922,7 +1001,7 @@ impl<T: TypeBounds> Expr<T> {
 
     /// Run a closure on this expression and every child, in pre-order.
     pub fn traverse<F>(&self, func: &mut F)
-        where F: FnMut(&Expr<T>) -> ()
+        where F: FnMut(&Expr) -> ()
     {
         func(self);
         for c in self.children() {
@@ -945,7 +1024,7 @@ impl<T: TypeBounds> Expr<T> {
 
     /// Recursively transforms an expression in place by running a function on it and optionally replacing it with another expression.
     pub fn transform_and_continue<F>(&mut self, func: &mut F)
-        where F: FnMut(&mut Expr<T>) -> (Option<Expr<T>>, bool)
+        where F: FnMut(&mut Expr) -> (Option<Expr>, bool)
     {
         match func(self) {
             (Some(e), true) => {
@@ -967,7 +1046,7 @@ impl<T: TypeBounds> Expr<T> {
     /// Recursively transforms an expression in place by running a function on it and optionally replacing it with another expression.
     /// Supports returning an error, which is treated as returning (None, false)
     pub fn transform_and_continue_res<F>(&mut self, func: &mut F)
-        where F: FnMut(&mut Expr<T>) -> WeldResult<(Option<Expr<T>>, bool)>
+        where F: FnMut(&mut Expr) -> WeldResult<(Option<Expr>, bool)>
         {
             if let Ok(result) = func(self) {
                 match result {
@@ -991,7 +1070,7 @@ impl<T: TypeBounds> Expr<T> {
 
     /// Recursively transforms an expression in place by running a function on it and optionally replacing it with another expression.
     pub fn transform<F>(&mut self, func: &mut F)
-        where F: FnMut(&mut Expr<T>) -> Option<Expr<T>>
+        where F: FnMut(&mut Expr) -> Option<Expr>
     {
         if let Some(e) = func(self) {
             *self = e;
@@ -1005,7 +1084,7 @@ impl<T: TypeBounds> Expr<T> {
     /// Recursively transforms an expression in place by running a function first on its children, then on the root
     /// expression itself; this can be more efficient than `transform` for some cases
     pub fn transform_up<F>(&mut self, func: &mut F)
-        where F: FnMut(&mut Expr<T>) -> Option<Expr<T>>
+        where F: FnMut(&mut Expr) -> Option<Expr>
     {
         for c in self.children_mut() {
             c.transform(func);
@@ -1016,7 +1095,7 @@ impl<T: TypeBounds> Expr<T> {
     }
 
     /// Returns true if this expressions contains `other`.
-    pub fn contains(&self, other: &Expr<T>) -> bool {
+    pub fn contains(&self, other: &Expr) -> bool {
         if *self == *other {
             return true;
         }
@@ -1027,4 +1106,13 @@ impl<T: TypeBounds> Expr<T> {
         }
         return false;
     }
+}
+
+/// Create a box containing an untyped expression of the given kind.
+pub fn expr_box(kind: ExprKind, annot: Annotations) -> Box<Expr> {
+    Box::new(Expr {
+                 ty: Type::Unknown,
+                 kind: kind,
+                 annotations: annot,
+             })
 }
