+++ weld/expr_hash.rs
@@ -34,7 +34,7 @@ impl fmt::Debug for ExprHash {
 impl ExprHash {
     /// Recurisvely computes signatures for an expression and each of its subexpressions.
     /// The precise symbol names defined within the expression are ignored.
-    fn from_expr<'a, T: TypeBounds>(&mut self, expr: &'a Expr<T>,
+    fn from_expr<'a>(&mut self, expr: &'a Expr,
                                     symbol_positions: &mut fnv::FnvHashMap<&'a Symbol, Vec<i32>>,
                                     max_id: &mut i32) -> WeldResult<()> {
         // In expressions that define new symbols, subexpressions will be handled in the match.
@@ -149,7 +149,7 @@ impl ExprHash {
     }
 
     /// Create a signature from an expression.
-    pub fn from<T:TypeBounds>(expr: &Expr<T>) -> WeldResult<ExprHash> {
+    pub fn from(expr: &Expr) -> WeldResult<ExprHash> {
         let mut sig = ExprHash { hasher: fnv::FnvHasher::default() };
         let mut symbol_positions = fnv::FnvHashMap::default();
         let mut max_id = 0;
