+++ weld/program.rs
@@ -4,13 +4,13 @@
 use std::vec::Vec;
 
 use super::ast::Symbol;
-use super::partial_types::*;
+use super::ast::Expr;
 
 #[derive(Clone, Debug, PartialEq)]
 pub struct Program {
     pub macros: Vec<Macro>,
     /// Program body -- this will likely be a Lambda, but not always.
-    pub body: PartialExpr,
+    pub body: Expr,
 }
 
 /// A macro we will substitute at compile time.
@@ -18,5 +18,5 @@ pub struct Program {
 pub struct Macro {
     pub name: Symbol,
     pub parameters: Vec<Symbol>,
-    pub body: PartialExpr,
+    pub body: Expr,
 }
