+++ weld/transforms/annotator.rs
@@ -6,7 +6,7 @@ use ast::ExprKind::*;
 /// Forces parallelization of for loops inside an `Iterate`. In the current LLVM code generator,
 /// this prevents for loop continuations from making recursive calls to the beginning of the
 /// iteration, which can cause stack exhaustion.
-pub fn force_iterate_parallel_fors(expr: &mut Expr<Type>) {
+pub fn force_iterate_parallel_fors(expr: &mut Expr) {
     expr.transform_and_continue(&mut |ref mut e| {
         match e.kind {
              Iterate { .. } => {
